pyproject.toml
[tool.poetry]
name = "modbusbackup"
version = "0.1.0"
description = "Backup connection between 2 unipi units using modbus"
authors = ["Martijn Hemeryck <martijn.hemeryck@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pymodbus = {version = "^2.5.2", extras = ["repl"]}
websockets = "^10.0"
requests = "^2.26.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
types-requests = "^2.25.11"
types-PyYAML = "^5.4.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
